chore: move inductee role commands to new arch
